+++ app/components/folio/console/links/control_bar_component.js
@@ -1,6 +1,7 @@
 window.Folio.Stimulus.register('f-c-links-control-bar', class extends window.Stimulus.Controller {
   static values = {
-    json: Boolean
+    json: Boolean,
+    absoluteUrls: { type: Boolean, default: false },
   }
 
   onAddClick (e) {
@@ -8,7 +9,7 @@ window.Folio.Stimulus.register('f-c-links-control-bar', class extends window.Sti
 
     this.element.dispatchEvent(new CustomEvent('f-c-input-form-group-url:edit', {
       bubbles: true,
-      detail: { json: this.jsonValue }
+      detail: { json: this.jsonValue, absoluteUrls: this.absoluteUrlsValue }
     }))
   }
 })