+++ app/inputs/url_json_input.rb
@@ -2,6 +2,6 @@
 
 class UrlJsonInput < SimpleForm::Inputs::StringInput
   def input(wrapper_options = nil)
-    register_url_input(json: true, wrapper_options:)
+    register_url_input(json: true, wrapper_options:, options:)
   end
 end